// contracts/generated/typescript/models/AdminUser.ts
export interface AdminUser {
    /**
     *
     * @type {string}
     * @memberof AdminUser
     */
    actorId: string;
    /**
     *
     * @type {string}
     * @memberof AdminUser
     */
    username: string;
    /**
     *
     * @type {string}
     * @memberof AdminUser
     */
    email: string;
    /**
     *
     * @type {string}
     * @memberof AdminUser
     */
    displayName: string;
    /**
     *
     * @type {AdminUserStatusEnum}
     * @memberof AdminUser
     */
    status: AdminUserStatusEnum;
    /**
     *
     * @type {Array<RoleSummary>}
     * @memberof AdminUser
     */
    roles: Array<RoleSummary>;
    /**
     *
     * @type {Date}
     * @memberof AdminUser
     */
    createdAt: Date;
}


/**
 * @export
 */
export const AdminUserStatusEnum = {
    Active: 'active',
    Disabled: 'disabled'
} as const;
export type AdminUserStatusEnum = typeof AdminUserStatusEnum[keyof typeof AdminUserStatusEnum];